[Allocator] Implement GPU EVAllocator.
