+++ tensorflow/core/framework/allocator_registry.h
@@ -44,6 +44,9 @@ class AllocatorFactory {
   //Create EV Allocator.
   virtual Allocator* CreateEVAllocator() {return CreateAllocator();};
 
+  // Create GPU EV Allocator.
+  virtual Allocator* CreateGPUEVAllocator() { return CreateAllocator(); }
+
   // Create a SubAllocator. If NumaEnabled() is true, then returned SubAllocator
   // will allocate memory local to numa_node.  If numa_node == kNUMANoAffinity
   // then allocated memory is not specific to any NUMA node.
@@ -85,6 +88,8 @@ class AllocatorFactoryRegistry {
 #endif
 
   Allocator* GetEVAllocator();
+
+  Allocator* GetGPUEVAllocator();
 
   // Returns 'best fit' SubAllocator.  First look for the highest priority
   // factory that is NUMA-enabled.  If none is registered, fall back to the
